requirements.txt
black==22.3.0
dataclasses==0.6
future==0.18.2
hopfield-layers==1.0.2
nltk==3.7
pip-chill==1.0.1
sklearn==0.0
torchtext==0.8.1
wandb==0.12.17
